# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统 API 文档",
        "title": "Kingfisher Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/departments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页部门列表；filter 支持 subtree_id（某部门及其子孙）、parent_id、status、q=名称搜索",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "部门列表（分页）",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词（搜索名称）",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "筛选 JSON，如 {\\",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-id",
                        "description": "排序",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "部门列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kingfisher_core_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/kingfisher_core_response.PageData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "创建部门",
                "parameters": [
                    {
                        "description": "创建请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/extends_department_transport.CreateDeptReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kingfisher_core_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/kingfisher_extends_department_domain.Department"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/departments/tree": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回全部部门树（含每个部门挂载的角色）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "部门树",
                "responses": {
                    "200": {
                        "description": "部门树",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kingfisher_core_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/kingfisher_extends_department_domain.Department"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/departments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回单个部门（含挂载的角色）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "部门详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "10901": {
                        "description": "部门不存在",
                        "schema": {
                            "$ref": "#/definitions/kingfisher_core_response.Response"
                        }
                    },
                    "200": {
                        "description": "部门详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kingfisher_core_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/kingfisher_extends_department_domain.Department"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "更新部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/extends_department_transport.UpdateDeptReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/kingfisher_core_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "有子部门时拒绝删除（返回 10902）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "10902": {
                        "description": "存在子部门，不可删除",
                        "schema": {
                            "$ref": "#/definitions/kingfisher_core_response.Response"
                        }
                    },
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/kingfisher_core_response.Response"
                        }
                    }
                }
            }
        },
        "/departments/{id}/roles": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "全量替换部门的角色关联（部门角色会合并进成员用户的有效权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "分配部门角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色ID列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/extends_department_transport.AssignRolesReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分配成功",
                        "schema": {
                            "$ref": "#/definitions/kingfisher_core_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "extends_department_transport.AssignRolesReq": {
            "type": "object",
            "required": [
                "role_ids"
            ],
            "properties": {
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "extends_department_transport.CreateDeptReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "技术部"
                },
                "parent_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "extends_department_transport.UpdateDeptReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "kingfisher_core_response.PageData": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "kingfisher_core_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "kingfisher_extends_department_domain.Department": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kingfisher_extends_department_domain.Department"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "description": "0=根",
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "role_ids": {
                    "description": "部门直接挂载的角色 ID",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "roles": {
                    "description": "部门挂载的角色详情",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kingfisher_extends_department_domain.Role"
                    }
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "description": "1=启用 0=停用",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "kingfisher_extends_department_domain.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}